Add Utils tests Add TestConf.cs file with public paths
