+++ CodeGen.Test/TestConf.cs
@@ -1,0 +1,10 @@
+﻿namespace CodeGen.Test
+{
+	public class TestConf
+	{
+		//CodeGen/CodeGen.Test/bin/Debug/netcoreapp2.0/CodeGen.Test.dll
+		public static readonly string BaseDir = "../../../../";
+		public static readonly string AssetsDir = "../../../testAssets";
+		
+	}
+}
